added update comment feature

+++ memo/views.py
@@ -4,9 +4,9 @@ from django.contrib.auth.mixins import LoginRequiredMixin, UserPassesTestMixin
 from django.views.generic import ListView, DetailView, CreateView, UpdateView, DeleteView
 from django.db import models
 from django.db.models import Q
-from .models import Post
+from .models import Post, Comment
 from django.http import HttpResponse, JsonResponse
-from django.views.decorators.http import require_POST
+from django.views.decorators.http import require_POST, require_http_methods
 from .forms import PostUpdateForm, PostCreateForm, CommentForm
 from django.contrib import messages
 
@@ -122,7 +122,6 @@ def create_comment(request):
             comment = form.save(commit=False)
             comment.author = request.user
             comment.save()
-            # messages.success(request, f'Comment has been added successfully.')
 
             # Prepare the data to be returned in the JSON response
             data = {
@@ -139,27 +138,39 @@ def create_comment(request):
     
     return redirect('post_detail', pk=comment.memo.pk)  # Redirect to the memo detail page
 
-def delete_comment(request, comment_id):
-    comment = get_object_or_404(comment, id=comment_id)
+@require_http_methods(['DELETE'])
+def delete_comment(request, pk):
+    comment = get_object_or_404(Comment, id=pk)
+    pk = comment.memo.pk
     if request.user == comment.author or request.user.is_superuser:
-        comment.delete()
-        messages.success(request, 'Comment has been deleted successfully.')
+        if request.method == 'DELETE':
+            comment.delete()
+            messages.success(request, 'Comment has been deleted successfully.')
+        else:
+            messages.error(request, 'Invalid request method.')
     else:
         messages.error(request, 'You are not allowed to delete this comment.')
-    return redirect('post_detail', pk=comment.memo.pk)
+    return redirect('post-detail', pk=pk)
 
-def update_comment(request, comment_id):
-    comment = get_object_or_404(comment, id=comment_id)
-    if request.user == comment.author:
-        if request.method == 'POST':
+def update_comment(request):
+    if request.method == 'POST':
+        comment = get_object_or_404(Comment, pk=int(request.POST.get('comment_id')))
+        if request.user == comment.author or request.user.is_superuser:
             form = CommentForm(request.POST, instance=comment)
             if form.is_valid():
                 form.save()
-                messages.success(request, 'Comment has been updated successfully.')
-                return redirect('post_detail', pk=comment.memo.pk)
-        else:
-            form = CommentForm(instance=comment)
-        return render(request, 'memo/update_comment.html', {'form': form, 'comment': comment})
-    else:
-        messages.error(request, 'You are not allowed to update this comment.')
-        return redirect('post_detail', pk=comment.memo.pk)
+                # Prepare the data to be returned in the JSON response
+                data = {
+                    'id': comment.pk,
+                    'content': comment.content,
+                    'date_posted': comment.date_posted.strftime('%Y-%m-%d %H:%M:%S'),
+                    'memo': comment.memo.pk,
+                }
+                # messages.success(request, 'Comment has been updated successfully.')
+                return JsonResponse(data)
+            else:
+                # Handle form validation errors
+                errors = form.errors.as_json()
+                return JsonResponse({'error': errors}, status=400)
+    
+    return redirect('post_detail', pk=comment.memo.pk)  # Redirect to the memo detail page
